feat(routes): create and edit an application's routes from the application
